requirements.txt
torch>=1.13
transformers>=4.29
numpy
soundfile
loguru
tqdm
pyyaml
pytest
